+++ packages/kit/test/apps/basics/src/routes/anchor/index.svelte
@@ -1,4 +1,5 @@
 <h1>Welcome to a test project</h1>
+<a id="scroll-anchor" href="#last-anchor-2">Bottom of this page</a>
 <a id="first-anchor" href="/anchor/anchor#go-to-element">Anchor demo (first)</a>
 <div>Spacing</div>
 <a id="second-anchor" href="/anchor/anchor#go-to-element">Anchor demo (second)</a>
@@ -7,6 +8,7 @@
 <div>Spacing</div>
 <a id="last-anchor" href="/anchor/anchor#go-to-element">Anchor demo (last)</a>
 <a id="last-anchor-2" href="/anchor/anchor">Anchor demo (last 2)</a>
+<a id="routing-page" href="/routing/hashes/target">Different page</a>
 
 <style>
 	:global(body) {