+++ sites/kit.svelte.dev/src/routes/index.svelte
@@ -7,7 +7,7 @@
 </script>
 
 <script>
-	import Machine from '$img/svelte-kit-machine.webp?w=1440;800&format=avif;webp;png&meta'
+	import Machine from '$img/svelte-kit-machine.webp?w=1440;800&format=avif;webp;png&meta';
 	import { Hero, Blurb } from '@sveltejs/site-kit';
 </script>
 
@@ -95,11 +95,11 @@ npm run dev -- --open
 
 <style>
 	:global(.hero-container:dir(rtl)) {
-		max-inline-size: 116rem;
+		max-width: 116rem;
 	}
 
 	pre {
-		block-size: 100%;
+		height: 100%;
 		display: flex;
 		flex-direction: column;
 		color: var(--second-text);
@@ -110,12 +110,12 @@ npm run dev -- --open
 	}
 
 	.blurb-shifter {
-		margin-block-start: calc(-10rem + var(--side-nav));
+		margin-top: calc(-10rem + var(--side-nav));
 	}
 
 	@media (min-width: 900px) {
 		.blurb-shifter {
-			margin-block-start: -12em;
+			margin-top: -12em;
 		}
 	}
 </style>