+++ packages/adapter-static/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-static
 
+## 1.0.0-next.28
+
+### Patch Changes
+
+- [breaking] remove `createIndexFiles` option, derive from `trailingSlash` instead ([#3801](https://github.com/sveltejs/kit/pull/3801))
+
+## 1.0.0-next.27
+
+### Patch Changes
+
+- `precompress` option also compress wasm files ([#3790](https://github.com/sveltejs/kit/pull/3790))
+
 ## 1.0.0-next.26
 
 ### Patch Changes