+++ packages/adapter-vercel/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sveltejs/adapter-vercel
 
+## 1.0.0-next.43
+
+### Patch Changes
+
+- Add prerendered routes to config ([#3822](https://github.com/sveltejs/kit/pull/3822))
+
+## 1.0.0-next.42
+
+### Patch Changes
+
+- Clean URLs and handle trailingSlash configuration ([#3820](https://github.com/sveltejs/kit/pull/3820))
+
+## 1.0.0-next.41
+
+### Patch Changes
+
+- update to Vite 2.8 and esbuild 0.14 ([#3791](https://github.com/sveltejs/kit/pull/3791))
+
 ## 1.0.0-next.40
 
 ### Patch Changes